requirements.txt
# Core ML/NLP
torch>=2.0.0
transformers>=4.45.0 
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Datasets
datasets>=2.16.0
huggingface-hub>=0.20.0

# Evaluation metrics
sacrebleu>=2.4.0
evaluate>=0.4.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=4.0.0
tqdm>=4.66.0
xlrd>=2.0.1
openpyxl>=3.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
hazm>=0.9.0  # Persian NLP toolkit
langdetect>=1.0.9
regex>=2023.0.0